Lay out the Telemetry board.
This is the initial draft of the Telemetry board layout. A few
debugging pins changed (on the SD card) for easier routing.

There is still some uncertainty with regard to the routing of the
Ethernet traces. According to my calculations, the differential
impedence between the RX+/- traces and the TX+/- traces is around
150R, which is 50% higher than one source was recommending. I am
uncertain how to reduce this though, and I am unclear as to how
much this will affect performance, especially when the traces
cover such a short distance.