Telemetry v2.0 (#2)
* Updates project name and submodule

* Adds new telemetry board schematics. Still need to place GPS and Ethernet and some main uC essentials

* Updates common component library

* Adds ethernet schematics

* Adds GPS module to common library

* Updates common submodule to include GPS

* Adds connects all modules except for LEDs in schematics

* Adds logic analyzer points and LEDs

* Initial layout of the Telemtry board.

The ethernet interface needs more work.

* Lay out the Telemetry board.

This is the initial draft of the Telemetry board layout. A few
debugging pins changed (on the SD card) for easier routing.

There is still some uncertainty with regard to the routing of the
Ethernet traces. According to my calculations, the differential
impedence between the RX+/- traces and the TX+/- traces is around
150R, which is 50% higher than one source was recommending. I am
uncertain how to reduce this though, and I am unclear as to how
much this will affect performance, especially when the traces
cover such a short distance.

* Makes mounting hole text invisible and updates title

* Fix routing near crystal.

* Makes mounting hole text invisible and updates title

Co-authored-by: Sijin Woo <sijin.woo@utexas.edu>
Co-authored-by: Chase Block <chase@chaseblock.com>